requirements.txt
beautifulsoup4==4.6.3
requests
pyyaml
Pillow